Day 2 solution
Lookup hashes and passing block to `sum` FTW! The important bit in this
one was codifying the win, lose, and draw key/value pairs into hashes.
It took me a while to realize this, but once I did, it all came together
without too much trouble. Along the way, I was reminded that `sum` can
take a block, which is pretty rad!