Day 5 solution
Lots or objects, lots of factory methods, and proof that regex is a
double edged sword!

I spent a good bit of time exploring immutability with this one, and
leanred that `[1, 2, 3].dup` doesn't do everything I thought it did! The
array is a new array, but the objects within are not!